backend/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
